use jest to run tests, remove yarn.lock use package-lock
