+++ .babelrc
@@ -1,5 +1,4 @@
 {
-
   "env": {
     "development": {
       "presets": [
