+++ package.json
@@ -14,8 +14,7 @@
     "prepublish": "npm run build",
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
-    "test": "karma start --single-run",
-    "test-watch": "karma start"
+    "test": "BABEL_ENV=development jest"
   },
   "repository": {
     "type": "git",
@@ -41,6 +40,7 @@
     "babel-cli": "^6.0.0",
     "babel-core": "^6.0.0",
     "babel-eslint": "^7.0.0",
+    "babel-jest": "^23.4.2",
     "babel-loader": "^6.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.23.0",
@@ -48,26 +48,12 @@
     "babel-preset-react": "^6.23.0",
     "babel-register": "^6.23.0",
     "babel-runtime": "^6.0.0",
-    "chai": "^3.5.0",
-    "es5-shim": "^4.5.0",
-    "es6-shim": "^0.35.3",
     "eslint": "^3.17.1",
     "eslint-config-defaults": "^9.0.0",
     "eslint-plugin-react": "^6.10.0",
-    "karma": "~0.13.9",
-    "karma-chai": "~0.1.0",
-    "karma-chrome-launcher": "~0.2.0",
-    "karma-cli": "~0.1.0",
-    "karma-mocha": "~0.2.0",
-    "karma-mocha-reporter": "~1.1.1",
-    "karma-phantomjs-launcher": "^0.2.1",
-    "karma-sourcemap-loader": "~0.3.5",
-    "karma-webpack": "~1.7.0",
-    "mocha": "~2.3.3",
-    "phantomjs": "^1.9.18",
+    "jest": "^23.5.0",
     "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "webpack": "~1.14.0"
+    "react-dom": "^16.4.2"
   },
   "dependencies": {
     "prop-types": "^15.5.0",
