+++ test/.eslintrc
@@ -3,7 +3,9 @@
     "mocha": true
   },
   "globals": {
-    "assert": true,
+    "expect": true,
+    "Promise": true,
+    "jest": true,
   },
   "rules": {
     "no-script-url": 1,
