+++ test/recaptcha-wrapper.spec.js
@@ -7,12 +7,13 @@ const WIDGET_ID = "someWidgetId";
 
 const grecaptchaMock = {
   render (node, options) {
-    assert.ok(node, options);
+    expect(node).toBeTruthy();
+    expect(options).toBeTruthy();
     return WIDGET_ID;
   },
 
   getResponse (widgetId) {
-    assert.equal(widgetId, WIDGET_ID);
+    expect(widgetId).toBe(WIDGET_ID);
     return VALUE;
   },
 };
@@ -25,13 +26,13 @@ describe("ReCAPTCHAWrapper", () => {
     delete window.grecaptcha;
   });
   it("should be wrapped properly", () => {
-    assert.equal(ReCAPTCHA.displayName, "AsyncScriptLoader(ReCAPTCHA)");
+    expect(ReCAPTCHA.displayName).toBe("AsyncScriptLoader(ReCAPTCHA)");
   });
   it("should proxy functions", () => {
     const ReCaptchaRef = React.createRef();
     ReactTestUtils.renderIntoDocument(
-      <ReCAPTCHA sitekey="xxx" ref={ReCaptchaRef} />,
+      <ReCAPTCHA sitekey="xxx" ref={ReCaptchaRef} onChange={jest.fn()} />,
     );
-    assert.equal(ReCaptchaRef.current.getValue(), VALUE);
+    expect(ReCaptchaRef.current.getValue()).toBe(VALUE);
   });
 });
